Move `State` auxiliary types to a new module
Keep only the main traits in the root module.
